Create move API

+++ opencell-admin/ejbs/src/main/java/org/meveo/admin/storage/StorageFactory.java
@@ -61,6 +61,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.OpenOption;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
@@ -1047,32 +1049,42 @@ public class StorageFactory {
      * @param destKey destination key of object.
      *
      */
-    public static void moveObject(String srcKey, String destKey) {
-        log.debug("move object from source key {} to destination key {}", srcKey, destKey);
-        // copy object from srckey to destKey
-        CopyObjectRequest copyObjRequest = CopyObjectRequest.builder()
-                .sourceBucket(bucketName).sourceKey(srcKey)
-                .destinationBucket(bucketName).destinationKey(destKey)
-                .build();
-
-        try {
-            s3FileSystem.getClient().copyObject(copyObjRequest);
+    public static void moveFileOrObject(String srcKey, String destKey) {
+        if (storageType.equals(NFS)) {
+            try {
+                Files.move(Paths.get(srcKey), Paths.get(destKey), StandardCopyOption.ATOMIC_MOVE);
+            } catch (IOException e) {
+                log.error("IOException while moving file : {}", e.getMessage());
+                e.printStackTrace();
+            }
         }
-        catch (NoSuchKeyException e) {
-            log.error("NoSuchKeyException while copying object in addExtension method : {}", e.getMessage());
-        }
+        else if (storageType.equalsIgnoreCase(S3)) {
+            log.debug("move object from source key {} to destination key {}", srcKey, destKey);
+            // copy object from srckey to destKey
+            CopyObjectRequest copyObjRequest = CopyObjectRequest.builder()
+                    .sourceBucket(bucketName).sourceKey(srcKey)
+                    .destinationBucket(bucketName).destinationKey(destKey)
+                    .build();
 
-        log.debug("delete old object at source key {}", srcKey);
-        // delete old object
-        DeleteObjectRequest deleteObjRequest = DeleteObjectRequest.builder()
-                .bucket(bucketName).key(srcKey)
-                .build();
+            try {
+                s3FileSystem.getClient().copyObject(copyObjRequest);
+            }
+            catch (NoSuchKeyException e) {
+                log.error("NoSuchKeyException while copying object in addExtension method : {}", e.getMessage());
+            }
 
-        try {
-            s3FileSystem.getClient().deleteObject(deleteObjRequest);
-        }
-        catch (NoSuchKeyException e) {
-            log.error("NoSuchKeyException while deleting object in addExtension method : {}", e.getMessage());
+            log.debug("delete old object at source key {}", srcKey);
+            // delete old object
+            DeleteObjectRequest deleteObjRequest = DeleteObjectRequest.builder()
+                    .bucket(bucketName).key(srcKey)
+                    .build();
+
+            try {
+                s3FileSystem.getClient().deleteObject(deleteObjRequest);
+            }
+            catch (NoSuchKeyException e) {
+                log.error("NoSuchKeyException while deleting object in addExtension method : {}", e.getMessage());
+            }
         }
     }
 
@@ -1093,7 +1105,7 @@ public class StorageFactory {
             String destKey = formatObjectKey(destFile.getPath());
             log.debug("rename key object in S3 bucket from source key {} to destination key {}", srcKey, destKey);
 
-            moveObject(srcKey, destKey);
+            moveFileOrObject(srcKey, destKey);
 
             return true;
         }
